# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitorban API",
        "description": "Monitorban REST API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/api/v1/faq": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "List FAQs",
                "operationId": "9cbb7ade00cd968a2053003964dff374",
                "responses": {
                    "200": {
                        "description": "FAQ list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/faq/store": {
            "post": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Create FAQ",
                "description": "CREATE: Add a new FAQ.",
                "operationId": "1a76eb282e2adbf2835023500c2fa408",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FaqRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "FAQ created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/faq/{id}": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Show public FAQ",
                "description": "SHOW: Get a single FAQ details.",
                "operationId": "45c60c204f569178884a14babd7a8aaf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ details"
                    },
                    "404": {
                        "description": "FAQ not found"
                    }
                }
            }
        },
        "/api/v1/faq/{id}": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Show FAQ",
                "operationId": "32d45108a354e181a45fb0d3d79b7e0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "FAQ not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Update FAQ",
                "description": "UPDATE: Edit an FAQ.",
                "operationId": "4d48fc3285c59c9558012b472d020cf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FaqRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "FAQ updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "FAQ not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Delete FAQ",
                "description": "DELETE: Remove an FAQ.",
                "operationId": "77f527675bf0339aada41c099ef2d8d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "FAQ not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FaqRequestData": {
                "title": "FAQ Request",
                "description": "FAQ request body data",
                "required": [
                    "question",
                    "answer"
                ],
                "properties": {
                    "question": {
                        "type": "string",
                        "example": "How do I reset my password?"
                    },
                    "answer": {
                        "type": "string",
                        "example": "Use the password reset option on the login page."
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}